learning Activity week 3 completed

+++ week03/Fractions/Fraction.cs
@@ -47,7 +47,7 @@ public class Fraction
 
     public double GetDecimalValue()
     {
-        double decimalNum = _top / _bottom;
+        double decimalNum = (double)_top / (double)_bottom;
         return decimalNum;
     }
